chdtapis/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.1
Django==5.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
idna==3.8
jwcrypto==1.5.6
mysql==0.0.3
mysqlclient==2.2.4
oauthlib==3.2.2
pycparser==2.22
PyMySQL==1.1.1
pytz==2024.1
requests==2.32.3
six==1.16.0
somepackage==1.2.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
